DerivedFunctionElement: add ThreadLocal cache for expression copies
Removes synchronization between threads that try to evaluate the same derived function.
However this increases memory consumption and is only effective when the same threads evaluate
the same derived functions often.

This change requires the ForkJoinPool not to be recreated for each evaluation step as the ThreadLocals
would be thrown away each time.