Engine: reset the interrupted flag
The Engine thread will be interrupted by hardInterrupt, e.g.,
on termination.

If the Engine is idle and sleeping, for example in processNextCommand
with `commandQueue.take()`, an InterruptedException will be thrown,
which clears the interrupted flag.
NOTE: "By convention, any method that exits by throwing an
InterruptedException clears interrupt status when it does so."

But if the Engine was not sleeping, that flag must be cleared manually
to indicate the interrupt has been handled.
Otherwise later code / plugins might be confused about it and
interpret it like an additional hardInterrupt, e.g., when doing some
blocking operation upon termination.