fixed concurrent modification exception
ClassLibrary.addEntry is called from compiler plugins, which modifies
the entries list, which would lead to an ConcurrentModificationException
in KernelBackend.writeEntriesToDisk.

Now the library entries are stored in a separate list, and after calling
KernelBackend.writeEntriesToDisk it is checked whether new library
entries are present which will then be written in the next iteration.