catch only Exceptions, not Throwables
In general, Throwables should not be handled by Java applications directly,
and especially Error should be left for the JVM.

Libraries are expected to throw Exceptions only and not Errors (for
situations that can be handled at all).

Our own Exceptions (CoreASMIssue, EngineError, ParserError, ...) extend
now RuntimeException and are therefore also caught with Exceptions.