ODTImporter: rework Exception handling
Swallowing the Exceptions in `parseXml` is not only bad practice,
but actually leads to NPEs later on in `process`, as the document
is not expected to be null.

For `ODTImporter.main(String[] args)`, some Exceptions are already expected
and handled - extend this with printing the stack trace here and a nice
message.

For `Specification.loadSpec`, wrap the more detailed
ParserConfigurationException and SAXException in a more general
IOException, which is already handled down the road.